rename file to respect fe util naming
